Refactor Todo command into CRUD
