+++ src/main/java/seedu/address/logic/commands/read/ListTodoCommand.java
@@ -1,8 +1,7 @@
-package seedu.address.logic.commands.todo;
+package seedu.address.logic.commands.read;
 
 import static seedu.address.logic.parser.CliSyntax.TODO_COMMAND_WORD;
 
-import seedu.address.logic.commands.read.ListCommand;
 import seedu.address.model.Model;
 import seedu.address.model.todo.Todo;
