+++ src/main/java/seedu/address/logic/parser/HelpCommandParser.java
@@ -32,18 +32,18 @@ import seedu.address.logic.commands.read.ListEventCommand;
 import seedu.address.logic.commands.update.RemoveContactFromEventCommand;
 import seedu.address.logic.commands.update.RemoveContactFromLogEventCommand;
 import seedu.address.logic.commands.update.RemoveTagFromEventCommand;
-import seedu.address.logic.commands.todo.AddContactToTodoCommand;
-import seedu.address.logic.commands.todo.AddTagToTodoCommand;
-import seedu.address.logic.commands.todo.AddTodoCommand;
-import seedu.address.logic.commands.todo.ClearTodoCommand;
-import seedu.address.logic.commands.todo.DeleteTodoCommand;
-import seedu.address.logic.commands.todo.DisplayTodoInformationCommand;
-import seedu.address.logic.commands.todo.FilterTodoCommand;
-import seedu.address.logic.commands.todo.ListTodoCommand;
-import seedu.address.logic.commands.todo.MarkTodoAsDoneCommand;
-import seedu.address.logic.commands.todo.MarkTodoAsNotDoneCommand;
-import seedu.address.logic.commands.todo.RemoveContactFromTodoCommand;
-import seedu.address.logic.commands.todo.RemoveTagFromTodoCommand;
+import seedu.address.logic.commands.update.AddContactToTodoCommand;
+import seedu.address.logic.commands.update.AddTagToTodoCommand;
+import seedu.address.logic.commands.create.AddTodoCommand;
+import seedu.address.logic.commands.delete.ClearTodoCommand;
+import seedu.address.logic.commands.delete.DeleteTodoCommand;
+import seedu.address.logic.commands.read.DisplayTodoInformationCommand;
+import seedu.address.logic.commands.read.FilterTodoCommand;
+import seedu.address.logic.commands.read.ListTodoCommand;
+import seedu.address.logic.commands.update.MarkTodoAsDoneCommand;
+import seedu.address.logic.commands.update.MarkTodoAsNotDoneCommand;
+import seedu.address.logic.commands.update.RemoveContactFromTodoCommand;
+import seedu.address.logic.commands.update.RemoveTagFromTodoCommand;
 
 /**
  * Parses input arguments and creates a new HelpCommand object.
