+++ src/main/java/seedu/address/logic/parser/todo/FilterTodoCommandParser.java
@@ -14,8 +14,8 @@ import java.util.List;
 import seedu.address.commons.core.Operator;
 import seedu.address.commons.core.Pair;
 import seedu.address.commons.core.index.Index;
-import seedu.address.logic.commands.todo.FilterTodoCommand;
-import seedu.address.logic.commands.todo.FilterTodoCommand.TodoPredicate;
+import seedu.address.logic.commands.read.FilterTodoCommand;
+import seedu.address.logic.commands.read.FilterTodoCommand.TodoPredicate;
 import seedu.address.logic.parser.ArgumentMultimap;
 import seedu.address.logic.parser.ArgumentTokenizer;
 import seedu.address.logic.parser.Parser;
