+++ src/main/java/seedu/address/logic/parser/todo/MarkTodoAsNotDoneCommandParser.java
@@ -1,7 +1,7 @@
 package seedu.address.logic.parser.todo;
 
 import seedu.address.commons.core.index.Index;
-import seedu.address.logic.commands.todo.MarkTodoAsNotDoneCommand;
+import seedu.address.logic.commands.update.MarkTodoAsNotDoneCommand;
 import seedu.address.logic.parser.Parser;
 import seedu.address.logic.parser.ParserUtil;
 import seedu.address.logic.parser.exceptions.ParseException;
@@ -23,4 +23,3 @@ public class MarkTodoAsNotDoneCommandParser implements Parser<MarkTodoAsNotDoneC
     }
 
 }
-
