+++ src/test/java/seedu/address/logic/parser/HelpCommandParserTest.java
@@ -11,7 +11,7 @@ import seedu.address.logic.commands.CommandResult;
 import seedu.address.logic.commands.HelpCommand;
 import seedu.address.logic.commands.create.AddContactCommand;
 import seedu.address.logic.commands.create.AddEventCommand;
-import seedu.address.logic.commands.todo.AddTodoCommand;
+import seedu.address.logic.commands.create.AddTodoCommand;
 import seedu.address.model.Model;
 import seedu.address.model.ModelManager;
