+++ src/test/java/seedu/address/logic/parser/todo/DeleteTodoCommandParserTest.java
@@ -7,7 +7,7 @@ import static seedu.address.testutil.TypicalIndexes.INDEX_FIRST;
 
 import org.junit.jupiter.api.Test;
 
-import seedu.address.logic.commands.todo.DeleteTodoCommand;
+import seedu.address.logic.commands.delete.DeleteTodoCommand;
 
 public class DeleteTodoCommandParserTest {
