+++ src/test/java/seedu/address/logic/parser/todo/DisplayTodoInfoCommandParserTest.java
@@ -7,7 +7,7 @@ import static seedu.address.testutil.TypicalIndexes.INDEX_FIRST;
 
 import org.junit.jupiter.api.Test;
 
-import seedu.address.logic.commands.todo.DisplayTodoInformationCommand;
+import seedu.address.logic.commands.read.DisplayTodoInformationCommand;
 
 public class DisplayTodoInfoCommandParserTest {
